+++ lib/exceptions.dart
@@ -1,0 +1,1 @@
+export 'src/factory/builder_not_found_in_factory.exception.dart';