+++ lib/src/factory/typed_factory/typed.factory.dart
@@ -1,0 +1,7 @@
+import 'package:zam_core/zam_core.dart' show Builder;
+
+import '../factory.dart';
+
+class TypedFactory<INSTANCE extends Object> extends Factory<Type, INSTANCE> {
+  const TypedFactory(Map<Type, Builder<INSTANCE>> builders) : super(builders);
+}