+++ lib/zam_factory_pattern.dart
@@ -1,0 +1,3 @@
+export 'package:zam_core/zam_core.dart' show Builder;
+
+export 'src/factory/_.index.dart';