+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.1
 
-  * Includes @component_1@.
-  * Includes @component_2@.
-  * Includes @component_3@.
+  * Includes the core classes `Factory` and `ModifiableFactory`.
+  * Includes `TypedFactory` that lets us use class names as keys.
+  * Includes `ModifiableFactoryMixin` for coverting any factory into a modifiable factory.
+  * Includes `BuilderNotFoundInFactoryException` which is thrown when a builder is not found in factory.